+++ backend/src/post/dto/update-post.dto.ts
@@ -19,5 +19,5 @@ export class UpdatePostDto {
     example: "2983-2zed2-z2d2n-nac38",
     description: "Post's image/images",
   })
-  readonly image: string | string[];
+  readonly image: string;
 }